requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
python-multipart
pydantic
openai
Pillow
